[shelly] Fix timeDuration handling in DTOs (#17689)

+++ bundles/org.openhab.binding.shelly/src/main/java/org/openhab/binding/shelly/internal/api2/Shelly2ApiJsonDTO.java
@@ -580,7 +580,7 @@ public class Shelly2ApiJsonDTO {
             @SerializedName("timer_started_at")
             public Double timerStartedAt;
             @SerializedName("timer_duration")
-            public Integer timerDuration;
+            public Double timerDuration;
         }
 
         public static class Shelly2DeviceStatusResult {
@@ -882,7 +882,7 @@ public class Shelly2ApiJsonDTO {
         @SerializedName("timer_started_at")
         public Double timerStartetAt;
         @SerializedName("timer_duration")
-        public Integer timerDuration;
+        public Double timerDuration;
         public Double apower;
         public Double voltage;
         public Double current;
